Api/requirements.txt
flask
flask_pymongo
bcrypt
PyJWT
dnspython
datetime